dockerfile for image removal
